Made TimeZoneFactory into an immutable singleton
The data is now initialized thread safely during construction, and it
is no longer possible to change the databases. This also means all
the accessors are now thread safe without any extra locking besides
the one required by instance().

The setters are still declared in order not to break the API, but
the implementations will not just print a warning on using deprecated
methods.