Add a possibility to define an expiration time for each cache item separately.
The time constant given while contructing a cache is used as a global
constant and also as an indication whether to use (non zero value) or
not to use (zero value) time expiration. Separate value for an cache item
overrides the global value only when global value is non zero value.

The time stored to the container is now stored as expiration time instead
of storing insertion time. This way we avoid some calculation and the
code is a bit cleaner.